Possible fix for veteran give experience
