+++ FlashPoint/src/core/serializer.py
@@ -38,6 +38,7 @@ from src.action_events.turn_events.resuscitate_victim_event import ResuscitateEv
 from src.action_events.turn_events.ride_vehicle_event import RideVehicleEvent
 from src.action_events.turn_events.stop_leading_victim_event import StopLeadingVictimEvent
 from src.action_events.vehicle_placed_event import VehiclePlacedEvent
+from src.action_events.veteran_give_experience_event import VeteranGiveExperienceEvent
 from src.models.game_board.door_model import DoorModel
 from src.models.game_board.game_board_model import GameBoardModel
 from src.models.game_board.null_model import NullModel
@@ -483,6 +484,11 @@ class JSONSerializer(object):
         return CrewChangeEvent(PlayerRoleEnum(payload['_role']['value']), payload['_player_index'])
 
     @staticmethod
+    def _deserialize_veteran_give_experience_event(payload: Dict) -> VeteranGiveExperienceEvent:
+        player: PlayerModel = JSONSerializer._deserialize_player(payload['c_player'])
+        return VeteranGiveExperienceEvent(player)
+
+    @staticmethod
     def deserialize(payload: Dict) -> object:
         """
         Grab an object and deserialize it.
@@ -600,6 +606,8 @@ class JSONSerializer(object):
             return JSONSerializer._deserialize_stop_command_event(payload)
         elif object_type == FireDeckGunEvent.__name__:
             return JSONSerializer._deserialize_fire_deck_gun_event(payload)
+        elif object_type == VeteranGiveExperienceEvent.__name__:
+            return JSONSerializer._deserialize_veteran_give_experience_event(payload)
 
         logger.warning(f"Could not deserialize object {object_type}, not of recognized type.")
